svc/nlb/sg/lb: ensure elbv2 is handling security group IDs on Create
Ensure the Security Group IDs is added on NLB load balancer creation.

Additionally, this is fixing the BYO SG update scenario by detecting the
replaced SG on CLB and delete it when it is owned by controller. The
same behavior will be implemented in the BYO SG scenario for NLB too.